Added new layouts
